refactor(builtins): move the twenty-four tree-walker-only builtins
$type, $zip, $spread, $lookup, $pad, $power, $shuffle, $exists, $error,
$assert, the URL and base64 codecs, the format/parse numerics and the
date/time functions existed in exactly one of the three dispatch paths.
That is why $type(x) works and $map(arr, $type) raises.

Nothing is reconciled here because there was no second implementation to
reconcile against -- these relocate unchanged. The reference suite is the
oracle that matters for this one: the date/time and number-formatting
functions are thin in the differential corpus and heavy in the suite.
